requirements.txt
langgraph>=0.2.0
openai>=1.0.0
yfinance>=0.2.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0 